Create GPS_3GFONA_GPS
This is the code from Adafruit to run the GPS module. 

Line 87 of the code states that the device does not have GPRS (General Packet Radio Service). 

More info on GPRS is here: http://www.liveviewgps.com/blog/what-is-gprs-and-how-is-it-used-with-gps-tracking-devices/
Without GPRS the FONA 3G will only achieve lock through GPS which is slower to find a signal, and slower to update.