ITEAD Arduino Code developed on 21/12/15
This code talks through the Arduino gateway and at this time registers on the network. GPRS still needs to be enabled as well as GPS, then lock with satellite for lat/long and sending to server.

The baud rate of the ITEAD needed to be set to 9600 as it was too high for the Arduino. This was done by communicating directly with the modem and changing the baud rate using AT commands. The only issue was the drivers only existed for Windows XP. A special laptop was set up with XP to install these drivers.